+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fix #4477 exposing LeaderElector.release to force an elector to give up the lease
 * Fix #4935: improve HTTP client implementation selection messages
 * Fix #4975: exposing scale operations for all Resources
+* Fix #5022: adding additional buffering to ExecWatchInputStream
 * Fix #4992: Optimize Quantity parsing to avoid regex overhead
 * Fix #4998: removing the internal usage of the Serialization yaml mapper